emulator working!

+++ src/stack.h
@@ -2,16 +2,17 @@
 #define STACK_H
 
 #include <string.h>
+#include <stdint.h>
 
 #define SIZE 16
 
 typedef struct {
-    int m[SIZE];
-    int c;
+    uint16_t m[SIZE];
+    uint8_t c;
 } Stack;
 
 void StackInit(Stack *s);
-void push(Stack *s, int v);
-int pop(Stack *s);
+void push(Stack *s, uint16_t v);
+uint16_t pop(Stack *s);
 
 #endif
